Rework pytest test discovery
