+++ test_suite/agtest
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+import sys
+import pytest
+
+pytest.main(['tests'] + sys.argv[1:])
