+++ test_suite/pytest.ini
@@ -1,4 +1,5 @@
 [pytest]
 log_cli=true
 log_cli_level=warn
-addopts=-s
+addopts=-s --strict
+python_files="*.py"
