+++ test_suite/requirements.txt
@@ -26,3 +26,4 @@ six==1.12.0
 toml==0.10.0
 transitions==0.6.9
 yarl==1.3.0
+pytest-sugar
